SSCS-9814 MYA Demo - Audio files don't playback in MYA application (#2444)
* SSCS-9814 MYA Demo - Audio files don't playback in MYA application
