+++ src/main/java/uk/gov/hmcts/reform/sscs/model/tya/HearingRecording.java
@@ -13,5 +13,4 @@ public class HearingRecording {
     private String fileName;
     private String fileType;
     private String documentUrl;
-    private String documentBinaryUrl;
 }
