+++ src/main/java/uk/gov/hmcts/reform/sscs/service/citizenrequest/CitizenRequestService.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.reform.sscs.service.citizenrequest;
 
 import static org.apache.commons.io.FilenameUtils.getExtension;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.*;
+import static uk.gov.hmcts.reform.sscs.util.DocumentUtil.stripUrl;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -144,8 +145,7 @@ public class CitizenRequestService {
                         .map(r -> HearingRecording.builder()
                                 .fileName(r.getValue().getDocumentFilename())
                                 .fileType(getExtension(r.getValue().getDocumentFilename()))
-                                .documentUrl(r.getValue().getDocumentUrl())
-                                .documentBinaryUrl(r.getValue().getDocumentBinaryUrl())
+                                .documentUrl(stripUrl(r.getValue().getDocumentBinaryUrl()))
                                 .build())
                         .collect(Collectors.toList()))
                 .build();
