+++ src/main/java/uk/gov/hmcts/reform/sscs/util/DocumentUtil.java
@@ -29,4 +29,15 @@ public class DocumentUtil {
     public static String userFriendlyName(String documentType) {
         return StringUtils.capitalize(StringUtils.join(Arrays.stream(StringUtils.splitByCharacterTypeCamelCase(documentType)).map(StringUtils::uncapitalize).toArray(String[]::new), " "));
     }
+
+    public static String stripUrl(String documentBinaryUrl) {
+        if (documentBinaryUrl != null) {
+            int frontIndex = documentBinaryUrl.indexOf("documents") + 10;
+            int backIndex = documentBinaryUrl.indexOf("/binary");
+            if (frontIndex >= 10 && backIndex > 0) {
+                return documentBinaryUrl.substring(frontIndex, backIndex);
+            }
+        }
+        return documentBinaryUrl;
+    }
 }
