+++ src/test/resources/tya/hearingRecordingMya.json
@@ -9,8 +9,7 @@
         {
           "fileName": "Final Town House 16 Jun 2021.mp3",
           "fileType": "mp3",
-          "documentUrl": "http://dm-store:5005/documents/1432689d-fd1c-4d57-aaa3-52d4da030840",
-          "documentBinaryUrl": "http://dm-store:5005/documents/1432689d-fd1c-4d57-aaa3-52d4da030840/binary"
+          "documentUrl": "1432689d-fd1c-4d57-aaa3-52d4da030840"
         }
       ]
     }
